Migrate test_scrub to pytest
